Move kdm pkg out of manegement controllers
