+++ pkg/rkenodeconfigserver/nodeserver.go
@@ -17,9 +17,9 @@ import (
 
 	"github.com/pkg/errors"
 	util "github.com/rancher/rancher/pkg/cluster"
-	kd "github.com/rancher/rancher/pkg/controllers/management/kontainerdrivermetadata"
 	v3 "github.com/rancher/rancher/pkg/generated/norman/management.cattle.io/v3"
 	"github.com/rancher/rancher/pkg/image"
+	kd "github.com/rancher/rancher/pkg/kontainerdrivermetadata"
 	"github.com/rancher/rancher/pkg/librke"
 	"github.com/rancher/rancher/pkg/rkeworker"
 	"github.com/rancher/rancher/pkg/settings"
